+++ stockfish_trainer/__init__.py
@@ -1,0 +1,5 @@
+"""Stockfish trainer package."""
+
+from .app import main
+
+__all__ = ["main"]
