+++ stockfish_trainer/__main__.py
@@ -1,0 +1,6 @@
+"""Allow running the package with `python -m stockfish_trainer`."""
+
+from .app import main
+
+if __name__ == "__main__":
+    main()
