+++ stockfish_trainer/utils.py
@@ -1,0 +1,4 @@
+"""Utility helpers for the Stockfish trainer application."""
+
+def clamp(v, vmin, vmax):
+    return max(vmin, min(v, vmax))
